+++ core/sr-primitives/src/generic/unchecked_mortal_extrinsic.rs
@@ -84,7 +84,7 @@ where
 {
 	/// NOTE: this transaction is not tipped i.e. the tip value will be `None`. It does not really
 	/// matter what the last generic is since it is always `None`.
-	type Checked = CheckedExtrinsic<AccountId, Index, Call, u64>;
+	type Checked = CheckedExtrinsic<AccountId, Index, Call, u32>;
 
 	fn check(self, context: &Context) -> Result<Self::Checked, &'static str> {
 		Ok(match self.signature {
@@ -236,7 +236,7 @@ mod tests {
 	const DUMMY_ACCOUNTID: u64 = 0;
 
 	type Ex = UncheckedMortalExtrinsic<u64, u64, Vec<u8>, TestSig>;
-	type CEx = CheckedExtrinsic<u64, u64, Vec<u8>, u64>;
+	type CEx = CheckedExtrinsic<u64, u64, Vec<u8>, u32>;
 
 	#[test]
 	fn unsigned_codec_should_work() {
